MinistryOfCoolWalks

Documentation for MinistryOfCoolWalks.

Constants

MinistryOfCoolWalks.DEFAULT_LANES_ONEWAY

MinistryOfCoolWalks.HIGHWAYS_NOT_OFFSET

MinistryOfCoolWalks.HIGHWAYS_OFFSET

MinistryOfCoolWalks.DEFAULT_LANES_ONEWAY — Constant

DEFAULT_LANES_ONEWAY

default number of lanes in one direction of the street, by highway type. Used when there is no data available in the tags.

MinistryOfCoolWalks.HIGHWAYS_NOT_OFFSET — Constant

HIGHWAYS_NOT_OFFSET

list of highways, which should not be offset, usually because they can allready considered the center of a bikepath/sidewalk/footpath...

MinistryOfCoolWalks.HIGHWAYS_OFFSET — Constant

HIGHWAYS_OFFSET

list of highways, which should be offset to the edge of the street.