Centerline Correction

Introduction

When loading shadow graph from OSM data, the street geometry represents, by convention, the center of the streets, which usually is not the place where we see pedestrians walking. To get a more accurate picture of how much shadow there is on the sidewalks, we need to offset the ways to where pedestrians would usually walk. This procedure, including estimating the width of the street is handled here.

Default values

As the datamodel of OSM allows for a certain variability in completeness of the mapped tags on each street, we need to assume some sensible default values for the width of streets, as well as for the types of highways which denote the centerline of streets vs pedestrian spaces. We have:

MinistryOfCoolWalks.DEFAULT_LANES_ONEWAY

MinistryOfCoolWalks.HIGHWAYS_OFFSET

MinistryOfCoolWalks.HIGHWAYS_NOT_OFFSET

If you encounter highways in your dataset which are not present, or would like to change the values given, you can update/add/... them as you need.

API

MinistryOfCoolWalks.DEFAULT_LANES_ONEWAY

MinistryOfCoolWalks.HIGHWAYS_NOT_OFFSET

MinistryOfCoolWalks.HIGHWAYS_OFFSET

MinistryOfCoolWalks.check_building_intersection

MinistryOfCoolWalks.correct_centerlines!

MinistryOfCoolWalks.guess_offset_distance

MinistryOfCoolWalks.node_directions

MinistryOfCoolWalks.offset_line

MinistryOfCoolWalks.DEFAULT_LANES_ONEWAY — Constant

DEFAULT_LANES_ONEWAY

default number of lanes in one direction of the street, by highway type. Used as a fallback when there is no data available in the tags.

MinistryOfCoolWalks.HIGHWAYS_NOT_OFFSET — Constant

HIGHWAYS_NOT_OFFSET

list of highways, which should not be offset, usually because they can allready considered the center of a bikepath/sidewalk/footpath...

MinistryOfCoolWalks.HIGHWAYS_OFFSET — Constant

HIGHWAYS_OFFSET

list of highways, which should be offset to the edge of the street.

MinistryOfCoolWalks.check_building_intersection — Method

check_building_intersection(building_tree, offset_linestring)

checks if the linestring offset_linestring interescts with any of the buildings saved in the building_tree, which is assumend to be an RTree of the same structure as generated by build_rtree. Returns the offending geometry, or an empty list, if there are no intersections.

MinistryOfCoolWalks.correct_centerlines! — Function

correct_centerlines!(g, buildings, assumed_lane_width=3.5, scale_factor=1.0)

offsets the centerlines of streets (edges in g) stored in the edge prop :eg_geometry_base, to the estimated edge of the street and stores the result in :sg_street_geometry.

Repeated application of this function deletes all edgeprops added after loading the graph, or the last application of correct_centerline, apart from [:sg_osm_id, :sg_tags, :sg_street_geometry, :sg_geometry_base, :sg_street_length, :sg_parsing_direction, :sg_helper].

The information available in the edgeprops :sg_tags and :sg_parsing_direction is used to estimate the width of the street. If it is not possible to find the offset using these props, the assumed_lane_width is used in conjunction with the gloabal dicts DEFAULT_LANES_ONEWAY, HIGHWAYS_OFFSET and HIGHWAYS_NOT_OFFSET, to figure out how far the edge should be offset. This guess is then multiplied by the scale_factor, to get the final distance by wich we then offset the line.

If the highway is in HIGHWAYS_NOT_OFFSET, it is not going to be moved, no matter the contents of its tags. For the full reasoning and implementations see the source of guess_offset_distance.

We check if the offset line does intersect more buildings than the original line, to make sure that the assumend foot/bike path does lead through a building. If there have new intersections arrisen, we retry the offsetting with 0.9, 0.8, 0.7... times the guessed offset, while checking and, if true breaking, whether the additional intersections vanish.

We also update the locations of the helper nodes, to reflect the offset lines, as well as the ":sgstreetlength" prop, to reflect the possible change in length.

MinistryOfCoolWalks.guess_offset_distance — Function

guess_offset_distance(g, edge::Edge, assumed_lane_width=3.5)

estimates the the distance an edge of graph g has to be offset. Uses the props of the edge, the assumed_lane_width, used as a fallback in case the information can not be found solely in the props, as well as the global constants of DEFAULT_LANES_ONEWAY, HIGHWAYS_OFFSET and HIGHWAYS_NOT_OFFSET. This function calls:

guess_offset_distance(edge_tags, parsing_direction, assumed_lane_width=3.5)

estimates the distance the edge with the tags edge_tags, which has been parse in the direction of parsing_direction. (That is, if we had to go forwards or backwards through the OSM way in order to generate the edgeometry for this linestring. This is nessecary, due to the existence of the reverseway tags and possible asymetries, where streets have more lanes in one direction, than in the other.)

MinistryOfCoolWalks.node_directions — Method

node_directions(x, y)

calculates the (scaled) direction in which the nodes given by 'x' and 'y' coordinates need to be offset, such that the connections between the nodes remain parallel to the original connections. Returns array of 2d vectors.

MinistryOfCoolWalks.offset_line — Method

offset_line(line, distance)

creates new ArchGDAL linestring where all segments are offset parallel to the original segment of line, with a distance of distance. Looking down the line (from the first segment to the second...), a positive distance moves the line to the right, a negative distance to the left. The line is expected to be in a projected coordinate system, which is going to be applied to the new, offset line as well. If, continious offsetting the length of a line segment where to reach a length of 0, the two adjacent points are automatically merged and offsetting is continued using the new configuration. Does account for self intersections created by the the endpoints crossing a line segment. In this case, only the closed part of the curve will be preserved.