RTree Building

Introduction

Just two little functions to build an RTree out of ArchGDAL polygons.

API

MinistryOfCoolWalks.build_rtree

MinistryOfCoolWalks.rect_from_geom

MinistryOfCoolWalks.build_rtree — Method

build_rtree(geo_arr)

builds SpatialIndexing.RTree{Float64, 2} from an array containing ArchGDAL.wkbPolygons. The value of an entry in the RTree is a named tuple with: (orig=original_geometry,prep=prepared_geometry). orig is just the original object, an element from geo_arr, where prep is the prepared geometry, derived from orig. The latter one can be used in a few ArchGDAL functions to get higher performance, for example in intersection testing, because relevant values get precomputed and cashed in the prepared geometry, rather than precomputed on every test.

MinistryOfCoolWalks.rect_from_geom — Method

rect_from_geom(geom)

builds SpatialIndexing.Rect with the extent of the geomerty geom.