Routing

Introduction

It might be interesting to look shortest paths in out network, using different weights for each edge, based on the length in shade and sun, as well as an external parameter, which essentially weights the length of sunny parts against the length of shaded parts along every edge. Therefore, we introduce our own Real subtype, together with a MetaGraphs.MetaWeights based Weight-Matrix, which should just work in every Graphs.jl algorithm, which takes a Weight Matrix.

ShadowWeight and conditions on addition comparison

To calculate shortest paths, we need a non negative edge weight, a less-than and an addition operation with a (or possibly multiple) zero elements.

we use:

felt_length(w) = (1-w.a) * w.shade + (1+w.a) * sun

and

real_length(w) = w.shade + w.sun

Since we want to be able to reconstruct the real_length from the number w which has been used in the shortest path with the felt_length function, we additionally need the addition with zero to be invariant under both length operations (this needs to hold for every operation you might want to do on the results from routing with a custom felt_length). That is:

felt_length(w + zero) == felt_length(w) => real_length(w + zero) == real_length(w)

we archieve this by constraining a in (-1.0, 1.0), rather than a in [-1.0, 1.0]. In this way, a zero is of length 0 under both operations.

Applying this constraint can be argued in multiple ways: (here in German. TODO: translate this...)

Das Problem betrachtend: dieses a skaliert die relative länge der schatten zu den sonnenwegen. Bedeutet in meiner Story fühlt sich Schatten (1-a)/(1+a) länger an als Sonne. Und für a = 1 bedeutet dass, das sich Schatten 0 mal länger anfühlt, und für a=-1 ist das ding nicht mal definiert, aber bedeutet sowas wie, das sich Schatten unendlich viel länger anfühlt als Sonne, was, gegeben die Problemstellung, irgendwie nicht wirklich Sinn ergibt. In beiden Fällen würde das bedeuten, das es dir egal ist wie weit du läufst, wenn du dich zwischen verschiedenen kanten mit nur sonne oder nur schatten entscheiden musst. Und das ist irgendwie nicht sinnvoll. (besser wäre es, grenzwerte lim a -> 1.0 und lim a -> -1.0 zu betrachten.)

Algorithmisch: wenn ich die wahl zwischen verschiedenen strecken die nur in der sonne oder nur im schatten sind, (die jeweils eine gefühlte länge von null haben), dann ist es von der implementation abhängig, welchen dieser wege ich nehme. Die Tatsächliche Länge die ich aus diesen pfaden ausrechnen kann, ist also nicht sehr aussagekräftig, weil ich nicht weiß welche der (sehr tatsächlich sehr unterschiedlich langen) strecken mit gefühlter länge der algorithmus genommen hat. (sprich: wenn viele kanten gefühlte länge null haben, dann gibt es viele “kürzeste wege”, und ich bekomme halt nur einen davon, und welcher ist random)

Mathematisch: für kürzesten pfad brauche ich die beiden funktionen (min, +), und + muss ein nullelement haben. Wenn ich -1 und 1 zulasse, dann gibt es aber für diese werte sehr viele nullelemente für die addition (unter der gefühlten länge), was erstmal kein problem ist, aber, wenn ich dann die reale länge ausrechne, dann sind die halt nicht mehr alle nullelemente. Und dann ist es wieder abhängig vom algorithmus, wie oft ich (gefühlte) nullen addiere, was ich für eine reale länge bekomme.

Und die dinge gehen alle weg, wenn ich einfach a in (-1.0, 1.0) fordere.

ShadowWeightsLight

faster, but less flexible version of a custom weight matrix. Used in the same way as ShadowWeights, but the resulting PathState only gives the distances in felt lengths. To get the lengths in shadow an in the sun, we need to reevaluate_distances with a user supplied weight matrix. (Usually weights(g)) will work here. This gives the lengths of the paths in real world lengths. If we need to calculate the distances in shade and sun, we can either reevaluate the distances at a different values of a and solve the different felt weights for shade and sun, or we simply reevaluate the distances twice, once with a weightmatrix where only the shadows are weights, and once where only the sunny parts are weights.

API

MinistryOfCoolWalks.ShadowWeight

MinistryOfCoolWalks.ShadowWeights

MinistryOfCoolWalks.ShadowWeights

MinistryOfCoolWalks.ShadowWeightsLight

MinistryOfCoolWalks.ShadowWeightsLight

Base.:+

Base.:<

Base.:==

Base.getindex

Base.getindex

Base.size

Base.size

Base.typemax

Base.zero

Graphs.johnson_shortest_paths

MinistryOfCoolWalks.felt_length

MinistryOfCoolWalks.get_path_length

MinistryOfCoolWalks.real_length

MinistryOfCoolWalks.reevaluate_distances

MinistryOfCoolWalks.reevaluate_distances_slow

MinistryOfCoolWalks.to_SimpleWeightedDiGraph

MinistryOfCoolWalks.ShadowWeight — Type

ShadowWeight(a::Float64, shade::Float64, sun::Float64) <: Real

Typ representing the weight on one edge.

a represents the preference for shadow or sun, where a==0.0 signifies indifference, a ∈ (0.0, 1.0) favours shaded edges,

and a ∈ (-1.0, 0.0) favours sunny edges. Value must be in (-1.0, 1.0), otherwise, an Error is thrown.

shade represents the (real world) length of the edge in shade. Has to be non-negative, otherwise, an Error is thrown.

sun represents the (real world) length of the edge in the sun. Has to be non-negative, otherwise, an Error is thrown.

if shade or sun is Inf, the other value has to be Inf as well, otherwise, an error is thrown.

This also means, that shade+sun=real_world_street_length.

unsafe_ShadowWeight(a, shade, sun)

unsafe constructor for ShadowWeight, does not validate the inputs, used internally when we know that all conditions are fulfilled.

MinistryOfCoolWalks.ShadowWeights — Type

ShadowWeights{T<:Integer,U<:Real} <: AbstractMatrix{ShadowWeight}

Abstract Matrix type of ShadowWeights, usable as weights in graph-algorithms.

MinistryOfCoolWalks.ShadowWeights — Method

ShadowWeights(a, full_weights::I, shadow_weights::I) where {T<:Integer,U<:Real,I<:MetaGraphs.MetaWeights{T,U}}

Base constructor for ShadowWeights. a has to be in (-1.0, 1.0), otherwise an error will be thrown. full_weights and shadow_weights are the full lengths of the edges and the length of these edges in shadow, respectively. Make sure that all(shadow_weights .<= full_weights) == truefor all edges which exist, otherwise, the results might not be what you expect. Constructor checks thatmaximum(shadowweights)<typemax(U)andmaximum(fullweights)<typemax(U)`. (TODO: Maybe there is a faster way of doing this?)

ShadowWeights(g::AbstractMetaGraph, a; shadow_source=:shadowed_length)

Constructs the ShadowWeights from a MetaGraph and the a value. (See the docs of ShadowWeight for an explanation of the parameter.)

Assumes that weightfield(g) encodes the full length of each edge. Additionally, it is possible to set the field from which the length of the shadows will be extracted. The default value is :shadowed_length.

MinistryOfCoolWalks.ShadowWeightsLight — Type

ShadowWeightsLight{T<:Integer,U<:Real} <: AbstractMatrix{U}

AbstractMatrix type, usable in graph algorithms, alternative approach to using ShadowWeights.

MinistryOfCoolWalks.ShadowWeightsLight — Method

ShadowWeightsLight(a, geom_weights::I, shadow_weights::I) where {T<:Integer,U<:Real,I<:MetaGraphs.MetaWeights{T,U}}

Base constructor for ShadowWeightsLight. a has to be in (-1.0, 1.0), otherwise an error will be thrown. geom_weights and shadow_weights are the full lengths of the edges and the length of these edges in shadow, respectively. Make sure that all(shadow_weights .<= geom_weights) == true, otherwise, the results might not be what you expect. Constructor checks that maximum(shadow_weights)<typemax(U) and maximum(full_weights)<typemax(U). (TODO: Maybe there is a faster way of doing this?)

ShadowWeightsLight(g::AbstractMetaGraph, a; shadow_source=:shadowed_length)

Constructs the ShadowWeightsLight from a MetaGraph and the a value. (See getindex(m::ShadowWeightsLight, ...) for an explanation of this parameter.)

Assumes that weightfield(g) encodes the full length of each edge. Additionally, it is possible to set the field from which the length of the shadows will be extracted. The default value is :shadowed_length.

Base.:+ — Method

+(a::ShadowWeight, b::ShadowWeight)

Two general ShadowWeights are addable, if their a fields match. The result is a new ShadowWeight with the same a value and the sum of the sun and shadow fields of both ShadowWeights. The return value is generated using unsafe_ShadowWeight due to performance considerations. Make sure that you only input valid ShadowWeights.

Special care has to be taken when adding values which identify with either zero or infinity. In this case, we ignore the condition of the a fields having to be the same and return just the appropriate input.

Base.:< — Method

<(a::ShadowWeight, b::ShadowWeight)

a ShadowWeight is less than another, if its felt_length is less than the one of the other.

Base.:== — Method

==(a::ShadowWeight, b::ShadowWeight)

Two ShadowWeights are the considered equal, if their felt_lengths are the same.

Base.getindex — Method

getindex(w::ShadowWeights, u::Integer, v::Integer)

Get the ShadowWeight at index u,v. The length in the sun is calculated as abs(full_length-shadow_length), to account for numerical deviations where the edge might be slightly shorter than the shadow covering it. If the length in the shade is systematically longer than the full edge, this will not Error, but fail silently. Since we check the maximum values on construction, we can use unsafe_ShadowWeight to create the return value.

Base.getindex — Method

getindex(w::ShadowWeightsLight{T,U}, u::Integer, v::Integer)::U where {T<:Integer} where {U<:Real}

Get the length of an edge from u to v in the felt_length, defined as:

(1 - w.a) * shadow_length + (1 + w.a) * sun_length. a represents the preference for shadow or sun, where a==0.0 signifies indifference, a ∈ (0.0, 1.0) favours shaded edges, and a ∈ (-1.0, 0.0) favours sunny edges.

Base.size — Method

size(x::ShadowWeightsLight)

size of ShadowWeightsLight is size of geometry weights contained.

Base.size — Method

size(x::ShadowWeights)

The size of a ShadowWeights is the size of the full_weights field.

Base.typemax — Method

typemax(x::ShadowWeight) = typemax(typeof(x))
typemax(::Type{ShadowWeight})

returns the maximum value associated with the ShadowWeight Real. Equivalent to ShadowWeight(0.0, Inf, Inf)

Base.zero — Method

zero(x::ShadowWeight)
zero(::Type{ShadowWeight})

returns the zero value associated with the ShadowWeight Real. Equivalent to ShadowWeight(0.0, 0.0, 0.0).

Graphs.johnson_shortest_paths — Method

Graphs.johnson_shortest_paths(g::AbstractGraph{U}, distmx::AbstractMatrix{T}) where {U<:Integer,T<:ShadowWeight}

version of johnson_shortest_paths for distmx with ShadowWeight as entries, since we can not subtract these. Converts the graph and weights to SimpleWeightedDiGraph, to speed up the calculation and abstract away the complexity. (In reality, this is just a bunch of dijkstra_shortest_paths, wrapped to return a JohnsonState).

MinistryOfCoolWalks.felt_length — Method

felt_length(w::ShadowWeight)

returns the felt length of a ShadowWeight. It is defined as: (1 - a) * shade + (1 + a) * sun

MinistryOfCoolWalks.get_path_length — Method

get_path_length(path, weights)

function to calculate the length of a path given by a vector of node ids in a externally supplied weight matrix. (Not exported, mainly used in Testing.)

MinistryOfCoolWalks.real_length — Method

real_length(w::ShadowWeight)

returns the real length of a ShadowWeight. (That is: sun+shade).

MinistryOfCoolWalks.reevaluate_distances — Method

reevaluate_distances(state, weights)

reevaluates the shortest paths in state with the given weight matrix. This function is neccessary since the approach with ShadowWeightsLight weights only returns the felt lengths. To make these values comparable, we need the path lengths under the same weight matrix. This algorithm works only for FloydWarshallStates, as it uses a modified floyd warshall algorithm to do so.

Please note that the results from this algorithm might vary from the results which can be obtained from other implementations of this reevaluation (mainly reevaluate_distances_slow), if there exist multiple shortest paths in the felt measure. In this, faster implementation, the one that ends up getting picked for the real length depends on the order in which the nodes are checked. But the routing output does so as well, so on average, we should be fine.

MinistryOfCoolWalks.reevaluate_distances_slow — Method

reevaluate_distances_slow(state, weights)

recalculates the lengths of the paths encoded in state using the supplied weights matrix. (Not exported, mainly used in Testing, since very slow.)

MinistryOfCoolWalks.to_SimpleWeightedDiGraph — Function

to_SimpleWeightedDiGraph(g, distmx)

converts graph g with weights in distmx into SimpleWeightedDiGraph with weights from distmx.