Measures based on routing on Graphs

Introduction

These are some helper function to calculate shortest paths and related measures on shadow graphs.

API

Graphs.johnson_shortest_paths

MinistryOfCoolWalks.betweenness_centralities

MinistryOfCoolWalks.early_stopping_dijkstra

MinistryOfCoolWalks.edges_visited

MinistryOfCoolWalks.to_SimpleWeightedDiGraph

Graphs.johnson_shortest_paths — Method

Graphs.johnson_shortest_paths(g::AbstractGraph{U}, distmx::AbstractMatrix{T}; max_length=typemax(T)) where {U<:Integer,T<:ShadowWeight}

version of johnson_shortest_paths for distmx with ShadowWeight as entries, since we can not subtract these. Converts the graph and weights to SimpleWeightedDiGraph (while making sure that zero length edges are not dropped (see to_SimpleWeightedDiGraph)), to speed up the calculation and abstract away the complexity. (In reality, this is just a bunch of early_stopping_dijkstra, wrapped to return a JohnsonState).

MinistryOfCoolWalks.betweenness_centralities — Method

betweenness_centralities(state, start)

Calculates the node and edge betweennesses encoded in the dijkstra state, with shortest paths from s. Assumes all paths to be unique (does not use to state.predecessors) field. Does not include the endpoints, and is not normalised. Most of this code was taken from the Graphs.jl betweenness_centrality implementation.

MinistryOfCoolWalks.early_stopping_dijkstra — Method

early_stopping_dijkstra(g::AbstractGraph, s::U, distmx::AbstractMatrix{T}=weights(g); max_length::T=typemax(T)) where {T<:Real,U<:Integer}

Calculates dijkstras shortest paths, but treats edges which would make the shortest path longer than max_length as non-existent.

Nearly all of this code was taken from the Graphs.jl dijkstra implementation.

Behaves like dijkstra_shortest_paths(g, i; trackvertices=true), if all other arguments are omitted.

MinistryOfCoolWalks.edges_visited — Method

edges_visited(parents::AbstractArray, reachables::BitArray)
edges_visited(state::Graphs.DijkstraState, reachables::BitArray)

Efficiently calculates all the edges traversed by the paths, considering only destinations i where reachables[i]==true.

MinistryOfCoolWalks.to_SimpleWeightedDiGraph — Function

to_SimpleWeightedDiGraph(g, distmx)

converts directed graph g with weights in distmx into SimpleWeightedDiGraph with weights from distmx. Due to the way we construct it here, edges with zero length are kept as structural nonzeros in the SWG.