Hexagonal Bins

Introduction

These utilities can be used to aggregate arbitrary spatial properties into hexagonal areas. Might be useful. Note that these functions are lacking tests.

API

MinistryOfCoolWalks.get_crs_for_hexagons

MinistryOfCoolWalks.hex_center

MinistryOfCoolWalks.hexagon_area

MinistryOfCoolWalks.hexagon_histogram

MinistryOfCoolWalks.hexagonify

MinistryOfCoolWalks.hexes2polys

MinistryOfCoolWalks.hexgrid_buffer

MinistryOfCoolWalks.get_crs_for_hexagons — Method

get_crs_for_hexagons(df::DataFrame)
get_crs_for_hexagons(g::AbstractMetaGraph)

unified interface for accessing the local coordinate system of dataframes and graphs that support it.

MinistryOfCoolWalks.hex_center — Method

hex_center(hex::Hexagons.Hexagon, r)

caculates the centerpoint of the hex with radius r. Returns an ArchGDAL point.

MinistryOfCoolWalks.hexagon_area — Method

hexagon_area(dx, dy)

area of a hexagon with radius 1, that is scaled by dx in x direction, and by dy in y direction.

MinistryOfCoolWalks.hexagon_histogram — Method

hexagon_histogram(aggregator, geom_source, radius; buffer=0, danger_value=10000, filter_values=x -> true)

calculates the "generalised histogram" for data in a DataFrame or MetaGraph. For both these types, we expect the following methods to be implemented:

project_local!

build_rtree (val should be a NamedTuple with at least the ArchGDAL.IGeometry under the keyword of orig).

hexagonify

get_crs_for_hexagons

project_back!

Arguments

aggregator: closure (actual_intersections::Vector{SpatialElem}, hexagon::ArchGDAL.IGemometry{WKBPolygon})->T, where T is the type of result in each hexagon-bin. the first Argument is a Vector of the SpatialElements in the R-Tree of geom_source, where ArchGDAL.intersects(hexagon, intersection.val.orig)==true. The second argument is the hexagon which is currently aggregated over.

geom_source: source of geometry for R-Tree (and maybe other data.), currently, DataFrame and AbstractMetaGraph are supported.

radius: radius of hexagons in which the bins should be calculated. (Passed to hexagonify).

Keyword Arguments

buffer=0: Number of times the hexagons should be buffered before the aggregation.

danger_value=10000: Number of predicted hexes above which hexagonify will fail.

filter_values=x->true: Only hexagons where the filter_values(aggregation_value)==true will be returned.

Process

We project geom_source to a local coordinate system and calculate the hexagonalisation of the area defined by it. The projected geom_source is then converted into an R-Tree. For each hexagon in the hexagonalisation we calculate the value within that hexagon using the aggregator, filter these values with filter_values, project everything back, and return only the hexagons and values where filter_values(value)==true.

Returns (filtered_hexagons, filtered_values)

Examples

To get the number of nodes of a ShadowGraph in each hexagon with more than 0 nodes, you can do something like this:

hexes, values = hexagon_histogram(shadow_graph, 50; filter_values=(>(0))) inters, hex
    count(i -> i.val.type == :vertex, inters)
end

To get the total area of (possibly overlapping) buildings do:

hexes, values = hexagon_histogram(buildings, 50) do inters, hex
    geom = mapreduce(ArchGDAL.union, inters; init=ArchGDAL.createpolygon()) do i
        ArchGDAL.intersection(hex, i.val.orig)
    end
    return ArchGDAL.geomarea(geom)
end

MinistryOfCoolWalks.hexagonify — Method

hexagonify(geometries, hex_radius; kwargs...)
hexagonify(df::DataFrame, hex_radius; kwargs...) = hexagonify(df.geometry, hex_radius; kwargs...)
hexagonify(g::AbstractMetaGraph, hex_radius; kwargs...)
hexagonify(polygon::ArchGDAL.IGeometry, hex_radius; buffer=0, danger_value=10000)

puts a bunch of hexagons with radius hex_radius in the area given by:

the convex hull of a vector of geometries.

the convex hull of the colum titled :geometry of a dataframe df.

the convex hull of the :sg_geometry prop in the vertices of (g).

a polygon.

All these things are expected to be ArchGDAL geometries.

A hexagon will be placed, if its centerpoint is within the polygon.

keyword arguments

buffer=0, number of times the resulting hex-grid should be buffered (that is, a layer of hexagons added to the outside) after filling the geometry.

danger_value=10000 if the approximated number of hexagons is larger than this value, we will throw an assertion error. This is to keep you from accidentally killing your REPL if you forget to convert between local and global corrdinate systems.

Returns a Vector of Hexagons.HexagonCubics.

MinistryOfCoolWalks.hexes2polys — Method

hexes2polys(hexes, hex_radius)

converts a vector of Hexangons.Hexagon with radius hex_radius to a vector of ArchGDAL polygons.

MinistryOfCoolWalks.hexgrid_buffer — Method

hexgrid_buffer(hexes)

returns a vector of all hexagons that touch at least one hexagon in hexes, without duplicates.