Routing

Introduction

It might be interesting to look shortest paths in out network, using different weights for each edge, based on the length in shade and sun, as well as an external parameter, which essentially weights the length of sunny parts against the length of shaded parts along every edge. Therefore, we introduce our own Real subtype, together with a MetaGraphs.MetaWeights based Weight-Matrix, which should just work in every Graphs.jl algorithm, which takes a Weight Matrix.

ShadowWeight and conditions on addition and comparison

To calculate shortest paths, we need a non negative edge weight, a less-than and an addition operation with a (or possibly multiple) zero elements.

we use:

felt_length(w) = w.shade + w.a * sun

and

real_length(w) = w.shade + w.sun

Since we want to be able to reconstruct the real_length from the number w which has been used in the shortest path with the felt_length function, we additionally need the addition with zero to be invariant under both length operations (this needs to hold for every operation you might want to do on the results from routing with a custom felt_length). That is:

felt_length(w + zero) == felt_length(w) => real_length(w + zero) == real_length(w)

we archieve this by constraining a in (0.0, Inf), rather than a in [0.0, Inf). In this way, a zero is of length 0 under both operations.

Our zero element takes the form of ShadowWeight(a, 0.0, 0.0) with an arbitrary a as constrained above.

API

MinistryOfCoolWalks.ShadowWeight

MinistryOfCoolWalks.ShadowWeights

MinistryOfCoolWalks.ShadowWeights

MinistryOfCoolWalks.SymmetricShadowWeights

Base.:*

Base.:*

Base.:*

Base.:+

Base.:<

Base.:==

Base.getindex

Base.size

Base.typemax

Base.zero

MinistryOfCoolWalks.felt_length

MinistryOfCoolWalks.get_path_length

MinistryOfCoolWalks.real_length

MinistryOfCoolWalks.reevaluate_distances

MinistryOfCoolWalks.reevaluate_distances_slow

MinistryOfCoolWalks.ShadowWeight — Type

ShadowWeight(a::Float64, shade::Float64, sun::Float64) <: Real

Type representing the weight on one edge.

a represents the preference for shadow or sun, where a==1.0 signifies indifference, a ∈ (1.0, Inf) favours shaded edges,

and a ∈ (0.0, 1.0) favours sunny edges. Value must be in (0.0, Inf), otherwise, an Error is thrown.

shade represents the (real world) length of the edge in shade. Has to be non-negative, otherwise, an Error is thrown.

sun represents the (real world) length of the edge in the sun. Has to be non-negative, otherwise, an Error is thrown.

if shade or sun is Inf, the other value has to be Inf as well, otherwise, an error is thrown.

This also means, that shade+sun=real_world_street_length.

unsafe_ShadowWeight(a, shade, sun)

unsafe constructor for ShadowWeight, does not validate the inputs, used internally when we know that all conditions are fulfilled.

MinistryOfCoolWalks.ShadowWeights — Type

ShadowWeights{T<:Integer,U<:Real} <: AbstractMatrix{ShadowWeight}

Abstract Matrix type with elements of ShadowWeights, usable as weights in graph-algorithms.

MinistryOfCoolWalks.ShadowWeights — Method

ShadowWeights(a, street_weights::I, shadow_weights::I) where {T<:Integer,U<:Real,I<:MetaGraphs.MetaWeights{T,U}}

Base constructor for ShadowWeights. a has to be in (0.0, Inf), otherwise an error will be thrown. street_weights and shadow_weights are the lengths of the streets and the lengths of shadow on these streets, respectively. Make sure that all(shadow_weights .<= street_weights) == true for all edges which exist, otherwise, the results might not be what you expect. Constructor checks that maximum(shadow_weights)<typemax(U) and maximum(street_weights)<typemax(U).

ShadowWeights(g::AbstractMetaGraph, a; shadow_source=:sg_shadow_length)

Constructs the ShadowWeights from a MetaGraph and the a value. (See the docs of ShadowWeight for an explanation of the parameter.)

Assumes that weightfield(g) encodes the full length of each street. Additionally, it is possible to set the field from which the length of the shadows will be extracted. The default value is :sg_shadow_length.

MinistryOfCoolWalks.SymmetricShadowWeights — Type

SymmetricShadowWeights{T<:Integer,U<:Real} <: AbstractMatrix{ShadowWeight}

Abstract Matrix type with elements of ShadowWeights, usable as weights in graph-algorithms. Always returns the smaller weight: distmx[i,j] = min(distmx[i,j], distmx[j,i]) if both directions exist. Otherwise, returns the existing one.

Base.:* — Method

*(a::AbstractFloat, b::ShadowWeight)
*(a::ShadowWeight, b::AbstractFloat)

Multiplication with Float. Scales shade and sun.

Base.:* — Method

*(a::Integer, b::ShadowWeight)
*(a::ShadowWeight, b::Integer)

Multiplication with Integer. Scales shade and sun.

Base.:* — Method

*(a::ShadowWeight, b::Bool)
*(a::Bool, b::ShadowWeight)

Multiplication with Boolean. Returns a if b is true, otherwise zero(ShadowWeight).

Base.:+ — Method

+(a::ShadowWeight, b::ShadowWeight)

Two general ShadowWeights are addable, if their a fields match. The result is a new ShadowWeight with the same a value and the sum of the sun and shadow fields of both ShadowWeights. The return value is generated using unsafe_ShadowWeight due to performance considerations. Make sure that you only input valid ShadowWeights.

Special care has to be taken when adding values which identify with either zero or infinity. In this case, we ignore the condition of the a fields having to be the same and return just the appropriate input.

Base.:< — Method

<(a::ShadowWeight, b::ShadowWeight)

a ShadowWeight is less than another, if its felt_length is less than the one of the other.

Base.:== — Method

==(a::ShadowWeight, b::ShadowWeight)

Two ShadowWeights are the considered equal, if their felt_lengths are the same.

Base.getindex — Method

getindex(w::ShadowWeights, u::Integer, v::Integer)

Get the ShadowWeight at index u,v. The length in the sun is calculated as abs(street_length-shadow_length), to account for numerical deviations where the edge might be slightly shorter than the shadow covering it. If the length in the shade is systematically longer than the full edge, this will not Error, but fail silently. Since we check the maximum values on construction, we can use unsafe_ShadowWeight to create the return value.

Base.size — Method

size(x::ShadowWeights)

The size of a ShadowWeights is the size of the street_weights field.

Base.typemax — Method

typemax(x::ShadowWeight) = typemax(typeof(x))
typemax(::Type{ShadowWeight})

returns the maximum value associated with the ShadowWeight Real. Equivalent to ShadowWeight(1.0, Inf, Inf)

Base.zero — Method

zero(x::ShadowWeight)
zero(::Type{ShadowWeight})

returns the zero value associated with the ShadowWeight Real. Equivalent to ShadowWeight(1.0, 0.0, 0.0).

MinistryOfCoolWalks.felt_length — Method

felt_length(w::ShadowWeight)

returns the felt length of a ShadowWeight. It is defined as: a * sun + shade

MinistryOfCoolWalks.get_path_length — Method

get_path_length(path, weights)

function to calculate the length of a path given by a vector of node ids in a externally supplied weight matrix. (Not exported, only used in Testing.)

MinistryOfCoolWalks.real_length — Method

real_length(w::ShadowWeight)

returns the real length of a ShadowWeight. (That is: sun+shade).

MinistryOfCoolWalks.reevaluate_distances — Method

reevaluate_distances(state, weights)

reevaluates the shortest paths in state with the given weight matrix. This function is used for testing. This algorithm works only for FloydWarshallStates, as it uses a modified floyd warshall algorithm to do so.

Please note that the results from this algorithm might vary from the results which can be obtained from other implementations of this reevaluation (mainly reevaluate_distances_slow), if there exist multiple shortest paths in the felt measure. In this, faster implementation, the one that ends up getting picked for the real length depends on the order in which the nodes are checked. But the routing output does so as well, so on average, we should be fine. (Not exported, only used in Testing.)

MinistryOfCoolWalks.reevaluate_distances_slow — Method

reevaluate_distances_slow(state, weights)

recalculates the lengths of the paths encoded in state using the supplied weights matrix. (Not exported, only used in Testing, since very slow.)