Shadow Itersection

Introduction

After moving around the ways, we now evaluate the effect of the shadows of the buildings, trees... on the ways in the network. Here we not only present the functionality to do so, but also a tool we build to clean up the result.

API

MinistryOfCoolWalks.add_shadow_intervals!

MinistryOfCoolWalks.combine_along_tree

MinistryOfCoolWalks.combine_lines

MinistryOfCoolWalks.get_length_by_buffering

MinistryOfCoolWalks.rebuild_lines

MinistryOfCoolWalks.add_shadow_intervals! — Method

add_shadow_intervals!(g, shadows)

adds the intersection of the polygons in dataframe shadows with metadata "center_lon" and "center_lat" and the geometry in the edgeprop :edgegeom of graph g to g. This operation can be repeated on the same graph with various shadows.

After this operation, all non-helper edges will have the additional property of ':shadowed_length'. This value is zero, if there is no shadow cast on the edge. If there is a shadow cast on the edge, the edge will have an additional property, ':shadowgeom', representing the geometry of the street in the shadow.

MinistryOfCoolWalks.combine_along_tree — Method

combine_along_tree(tree, start_node, lines, min_dist)

recursively combines the lines at leafs in tree with the nodes one order up, starting (as in, the recursion starts here. This node gets combined last) at the root start_node. The min_dist is needed to figure out how the lines should be combined. (This dependency could maybe be removed...)

MinistryOfCoolWalks.combine_lines — Method

combine_lines(a, b, min_dist)

combines two lines a and b at the ends where they are closer than min_dist apart.

If a ⊂ b, returns b, if b ⊂ a returns a. Otherwise, returns all nodes of a, concatenated with all nodes of b, which are further away from a than min_dist.

MinistryOfCoolWalks.get_length_by_buffering — Method

get_length_by_buffering(geom, buffer, points, edge)

approximates the "non overlapping" length of overlapping linestrings by buffering geom with a buffer of size buffer and points points, specifying how many points should be used to approximate bends of 90 degrees. After buffering, we calculate the area and divide by the twice the buffer width (buffer is radius), after subtracting the area of the endcaps. (This function is currently not in use.)

MinistryOfCoolWalks.rebuild_lines — Method

rebuild_lines(line::ArchGDAL.IGeometry{ArchGDAL.wkbLineString}, min_dist)

calculates the union of lines in a (multi) linestring, merging lines which are closer than min_dist to one another. This particular function just returns line, as a single line allready is the union

rebuild_lines(lines::ArchGDAL.IGeometry{ArchGDAL.wkbMultiLineString}, min_dist)::EdgeGeomType
rebuild_lines(lines, min_dist)::EdgeGeomType

We calculate the adjacency matrix of all lines, build a network with edges where the distance between edges < min_dist, calculate a dfs tree for each connected component and recursively combine the linestrings at the leafs with the linestrings one level up.