Shadow Itersection

Introduction

After moving around the ways, we now evaluate the effect of the shadows of the buildings, trees... on the ways in the network. The main function intended for use here is MinistryOfCoolWalks.add_shadow_intervals!. To check for possible errors we provide MinistryOfCoolWalks.check_shadow_angle_integrity

API

MinistryOfCoolWalks.add_shadow_intervals!

MinistryOfCoolWalks.all_less_than

MinistryOfCoolWalks.angles_in

MinistryOfCoolWalks.check_shadow_angle_integrity

MinistryOfCoolWalks.combine_along_tree

MinistryOfCoolWalks.combine_lines

MinistryOfCoolWalks.join_shadow_without_union!

MinistryOfCoolWalks.rebuild_lines

MinistryOfCoolWalks.add_shadow_intervals! — Method

add_shadow_intervals!(g, shadows; clear_old_shadows=false)

adds the intersection of the polygons in the :geometry column of the in dataframe shadows and the geometry in the edgeprop :sg_street_geometry of graph g to g.

This operation can be repeated on the same graph with various shadows.

If clear_old_shadows is true, all possible, preexisting effects of previous executions of this function are reset. This way, a once loaded graph can be reused for multiple experiments.

After this operation, all non-helper edges will have the additional property of :sg_shadow_length. This value is zero, if there is no shadow cast on the edge. If there is a shadow cast on the edge, the edge will have an additional property, :sg_shadow_geometry, representing the geometry of the street in the shadow.

MinistryOfCoolWalks.all_less_than — Method

all_less_than(angles, max_angle)

checks if all values in angles are less than max_angle.

MinistryOfCoolWalks.angles_in — Method

angles_in(line)
angles_in(::MultiLineStringTrait, lines)
angles_in(::LineStringTrait, line)

calculates all angles between segments in line. Result in radians.

MinistryOfCoolWalks.check_shadow_angle_integrity — Method

check_shadow_angle_integrity(g, max_angle)

checks, if all angles in shadows (:sg_shadow_geometry) in g are less than max_angle (in radians). If not, prints a warning. Used to test if the shadow joining works as intended.

MinistryOfCoolWalks.combine_along_tree — Method

combine_along_tree(tree, start_node, lines, min_dist)

recursively combines the lines at leafs in tree with the nodes one order up, starting (as in, the recursion starts here. This node gets combined last) at the root start_node. The min_dist is needed to figure out how the lines should be combined. (This dependency could maybe be removed...)

MinistryOfCoolWalks.combine_lines — Method

combine_lines(a, b, min_dist)

combines two lines a and b at the ends where they are closer than min_dist apart. This assumes, that a and b are (much) longer than min_dist. If not, weird edge cases may arrise.

If a ⊂ b, returns b, if b ⊂ a returns a. Otherwise, returns all nodes of a, concatenated with all nodes of b, which are further away from a than min_dist. If a and b form a circle, special care is taken to not mess up the order.

MinistryOfCoolWalks.join_shadow_without_union! — Method

join_shadow_without_union!(full_shadow, new_shadow::ArchGDAL.IGeometry{ArchGDAL.wkbLineString})
join_shadow_without_union!(full_shadow, new_shadow::ArchGDAL.IGeometry{ArchGDAL.wkbMultiLineString})

adds new_shadow to full_shadow, skipping empty linestrings.

MinistryOfCoolWalks.rebuild_lines — Method

rebuild_lines(lines::ArchGDAL.IGeometry{ArchGDAL.wkbMultiLineString}, min_dist)::EdgeGeomType
rebuild_lines(lines, min_dist)::EdgeGeomType

calculates the union of lines in a (multi) linestring, merging lines which are closer than min_dist to one another.

We calculate the adjacency matrix of all lines, build a network with edges where the distance between edges < min_dist, calculate a dfs tree for each connected component and recursively combine the linestrings at the leafs with the linestrings one level up.